use subdirectory as root for nginx example
This makes the exmaple a bit more realistic, because you wouldn't
usually want to serve the directory the .nix files are located in.

It also demonstrates how to point to a directory. Beginners might
struggle with trailing slashes and other details of specifying a path.

Finally it makes the example work as-is.